Move run_vega to scripts folder
